Update CareerBuilder API version from 1 to 2
Version 3 of the CareerBuilder API is being rolled out right now in internal CareerBuilder projects, so it's safe to assume v1 will be deprecated soon. Since there are no documented differences between v1 and v2 of the Job APIs, this commit updates the CB API endpoint from v1 to v2.